update by discussion
