+++ src/renderer/components/atoms/button/index.tsx
@@ -1,1 +1,3 @@
-export { default } from './button';
+/* eslint-disable import/prefer-default-export */
+export { default as Button } from './button';
+export { default as AnchorButton } from './anchor_button';
