+++ src/renderer/components/atoms/index.ts
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
 export { default as Text } from './text';
 export { default as Input } from './input';
-export { default as Button } from './button';
+export * from './button';
